feat(*): Refactoring

+++ employee-car-service/src/main/kotlin/com/vvalitsky/repository/EmployeeCarRepository.kt
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository
  *
  * @author Vladislav Valitsky
  */
-interface EmployeeCarRepository : JpaRepository<EmployeeCar, Long> {
-    fun findByEmployeeId(employeeId: Long): List<EmployeeCar>
+interface EmployeeCarRepository : JpaRepository<EmployeeCar, Int> {
+    fun findByEmployeeId(employeeId: Int): List<EmployeeCar>
 }
